Rename script for consistency across all workers scripts.
